Few Changes in Airport

+++ src/flight/Airport.java
@@ -2,10 +2,10 @@ package flight;
 
 public class Airport {
 	
-	private Float MAX_LON = (float) 90.0;
-	private Float MIN_LON = -90.0;
-	private Float MAX_LAT = 180.0;
-	private Float MIN_LAT = -180.0;
+	private double MAX_LON = 90.0;
+	private double MIN_LON = -90.0;
+	private double MAX_LAT = 180.0;
+	private double MIN_LAT = -180.0;
 	
 	private String name;
 	private Float latitude;
